Add Notification Service to Send Emails with Kafka
- Integrated Kafka into `docker-compose` for message brokering.
- Added Schema Registry to manage Avro schemas.
- Included Kafka UI for visualizing Kafka messages.
- Created Avro schema for message serialization.

+++ order-service/src/main/java/org/projects/orderservice/dto/InstrumentServiceResponseDto.java
@@ -2,8 +2,6 @@ package org.projects.orderservice.dto;
 
 import lombok.Builder;
 
-import java.math.BigDecimal;
-
 @Builder
 public record InstrumentServiceResponseDto(
         String id,
@@ -12,7 +10,7 @@ public record InstrumentServiceResponseDto(
 
         String title,
 
-        BigDecimal price,
+        Double price,
 
         Integer quantity
 ) {
