+++ order-service/src/main/java/org/projects/orderservice/model/InstrumentOrder.java
@@ -1,10 +1,11 @@
 package org.projects.orderservice.model;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
-
-import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -15,7 +16,7 @@ public class InstrumentOrder {
     @Id
     private Long id;
 
-    private BigDecimal price;
+    private Double price;
 
     private Integer quantity;
